Support non-splittable files in flag maker
The flag maker generates files based on a max flags value, which is
compared against an input file number of maps. The number of maps for an
input file is computed based on the block size of the file in HDFS.
However, this makes the assumption that the file is splittable/the
designated input format class can split files. If this is not the case
then calculations will be off. This change adds a user-settable flag
per datatype to indicate whether or not the input files are splittable
(with a default of true). If input files are not splittable, then the
number of maps for an input file is calculated as 1 to reflect that the
file will be referenced by only a single mapper on an ingest job.